requirements.txt
appdirs==1.4.4
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
jsonschema==4.17.3
makefun==1.15.1
mypy==1.4.0
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.2.0
pyee==8.2.2
pyppeteer==1.0.2
pyrsistent==0.19.3
pytest==7.3.2
pytest-steps==1.8.0
python-dotenv==1.0.0
refurb==1.16.0
requests==2.31.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
websockets==10.4
wrapt==1.15.0
zipp==3.15.0
